refactor: split filesystem.py into _fs_projects, _fs_reports, and _fs_clone
Extract project CRUD, report reading, and git clone logic from the
monolithic FilesystemActionProvider into focused module-level functions.
The class remains as a thin facade that delegates to the new modules.
